bfsmacklib: Altered play_smk_via_buffer() to use surfaces
Instead of raw buffer, the function now uses screen surface,
and then blits it to actual screen.

This future proofs the implementation, for the case we will
use accelerated GFX API at some point.

The implementation of the screen surface can be altered to
allow for buffer migration between RAM and VRAM. We should
move towards all pixel data buffers which are blitted onto
screen, to be allocated as part of surfaces.