syndwarsfx: Do not use optimized sprite draw in draw_phwoar()
Instead of optimized function, use the unified general function.
This makes it slower, but also allows linear scaling of the
explosion sprites, and restores draw flags support, like
transparency in final franes of explosion.

The speed decrease is not too drastic, and shouldn't be visible
on todays CPUs, unless a very high resolution is used.